+++ elrond-wasm-debug/src/api/managed_types/mod.rs
@@ -3,4 +3,4 @@ mod big_int_util;
 mod elliptic_curve_api_mock;
 mod managed_buffer_api_mock;
 mod managed_type_api_mock;
-mod static_buffer_api_mock;
+mod static_var_api_mock;
