+++ elrond-wasm-node/src/api/managed_types/mod.rs
@@ -2,4 +2,4 @@ pub(crate) mod big_int_api_node;
 mod elliptic_curve_api_node;
 pub(crate) mod managed_buffer_api_node;
 mod managed_type_api_node;
-mod static_buffer_api_node;
+mod static_var_api_node;
