+++ elrond-wasm-node/src/api/managed_types/static_var_api_node.rs
