+++ elrond-wasm/src/api/uncallable/mod.rs
@@ -13,7 +13,7 @@ mod managed_buffer_api_uncallable;
 mod managed_type_api_uncallable;
 mod print_api_uncallable;
 mod send_api_uncallable;
-mod static_buffer_api_uncallable;
+mod static_var_api_uncallable;
 mod storage_api_uncallable;
 mod vm_api_uncallable;
