+++ elrond-wasm/src/api/uncallable/static_var_api_uncallable.rs
